requirements.txt
streamlit==1.25.0
tensorflow==2.17.0
Pillow==9.4.0
numpy==1.24.0
joblib
scikit-learn